+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Unreleased
+## 0.27.0
+
+### Features
+
+- Add C++ exception monitor option ([#554](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/554))
+  - **Compose Multiplatform**: It is highly recommended to disable C++ monitoring (`options.enableUnhandledCppExceptionMonitoring = false`) on Apple targets. Unhandled Kotlin exceptions from CMP can be reported as generic C++ crashes (`ExceptionObjHolderImpl`) instead of useful Kotlin stack traces when C++ monitoring is enabled.
 
 ### Dependencies